requirements.txt
streamlit
google-generativeai
faiss-cpu
numpy
pandas
tqdm
sentence-transformers
python-dotenv
PyPDF2